Additions/solaris/{Installer,vboxms,vboxservice}: Add support for the Solaris Guest Additions to be installed into an alternate root path ('pkgad -R').  This also enables the Solaris GAs to be installed in an automated fashion when building Solaris images using the distribution constructor (aka the distro_const(8) utility).
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@101937 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/solaris/Installer/preremove.sh
@@ -44,18 +44,23 @@ export LANG
 echo "Removing VirtualBox service..."
 
 # stop and unregister VBoxService
-/usr/sbin/svcadm disable -s virtualbox/vboxservice
-# Don't need to delete, taken care of by the manifest action
-# /usr/sbin/svccfg delete svc:/application/virtualbox/vboxservice:default
-/usr/sbin/svcadm restart svc:/system/manifest-import:default
+if [ "${PKG_INSTALL_ROOT:-/}" = "/" ] ; then
+    /usr/sbin/svcadm disable -s svc:/application/virtualbox/vboxservice:default
+    /usr/sbin/svcadm disable -s svc:/application/virtualbox/vboxmslnk:default
+    # Don't need to delete, taken care of by the manifest action
+    #/usr/sbin/svccfg delete svc:/application/virtualbox/vboxservice:default
+    #/usr/sbin/svccfg delete svc:/application/virtualbox/vboxmslnk:default
+    /usr/sbin/svcadm restart -s svc:/system/manifest-import:default
 
-# stop VBoxClient
-pkill -INT VBoxClient
+    # stop VBoxClient
+    pkill -INT VBoxClient
+fi
 
 echo "Removing VirtualBox kernel modules..."
 
 # vboxguest.sh would've been installed, we just need to call it.
-/opt/VirtualBoxAdditions/vboxguest.sh stopall silentunload
+
+${PKG_INSTALL_ROOT}/opt/VirtualBoxAdditions/vboxguest.sh stopall silentunload
 
 # Figure out group to use for /etc/devlink.tab (before Solaris 11 SRU6
 # it was always using group sys)
@@ -71,22 +76,22 @@ if [ -f /etc/dev/reserved_devnames ]; then
 fi
 
 # remove devlink.tab entry for vboxguest
-sed -e '/name=vboxguest/d' /etc/devlink.tab > /etc/devlink.vbox
-chmod 0644 /etc/devlink.vbox
-chown root:$group /etc/devlink.vbox
-mv -f /etc/devlink.vbox /etc/devlink.tab
+sed -e '/name=vboxguest/d' ${PKG_INSTALL_ROOT}/etc/devlink.tab > ${PKG_INSTALL_ROOT}/etc/devlink.vbox
+chmod 0644 ${PKG_INSTALL_ROOT}/etc/devlink.vbox
+chown root:$group ${PKG_INSTALL_ROOT}/etc/devlink.vbox
+mv -f ${PKG_INSTALL_ROOT}/etc/devlink.vbox ${PKG_INSTALL_ROOT}/etc/devlink.tab
 
 # remove the link
-if test -h "/dev/vboxguest" || test -f "/dev/vboxguest"; then
-    rm -f /dev/vboxdrv
+if test -h "${PKG_INSTALL_ROOT}/dev/vboxguest" || test -f "${PKG_INSTALL_ROOT}/dev/vboxguest"; then
+    rm -f ${PKG_INSTALL_ROOT}/dev/vboxdrv
 fi
-if test -h "/dev/vboxms" || test -f "/dev/vboxms"; then
-    rm -f /dev/vboxms
+if test -h "${PKG_INSTALL_ROOT}/dev/vboxms" || test -f "${PKG_INSTALL_ROOT}/dev/vboxms"; then
+    rm -f ${PKG_INSTALL_ROOT}/dev/vboxms
 fi
 
 # Try and restore xorg.conf!
 echo "Restoring X.Org..."
-/opt/VirtualBoxAdditions/x11restore.pl
+${PKG_INSTALL_ROOT}/opt/VirtualBoxAdditions/x11restore.pl
 
 
 echo "Done."
